final sync: replace remote with local version
