+++ config/topo/topo_5.yaml
@@ -1,0 +1,37 @@
+version: 1
+topology:
+  type: MultiIf
+  subnets:
+    left: 10.0
+    right: 10.1
+
+  paths:
+    - link_type: c2r
+      id: 0
+      delay_ms: 100
+      queue_pkts: 20
+      bw_mbit: 4
+
+    - link_type: c2r
+      id: 1
+      delay_ms: 100
+      queue_pkts: 20
+      bw_mbit: 4
+
+    - link_type: r2s
+      id: 0
+      delay_ms: 10
+      queue_pkts: 20
+      bw_mbit: 10
+
+    - link_type: r2s
+      id: 1
+      delay_ms: 10
+      queue_pkts: 20
+      bw_mbit: 10
+
+    - link_type: r2s
+      id: 2
+      delay_ms: 10
+      queue_pkts: 20
+      bw_mbit: 10
