--- config/topo/topo_cong
@@ -1,5 +1,0 @@
-leftSubnet:10.0.
-rightSubnet:10.1.
-path_c2r_0:10,10,4
-path_c2r_1:40,30,4
-topoType:MultiIfMultiClient
